Design Patterns: Chain of Responsibility.

+++ DesignPatterns/src/main/java/behavioral/chain_of_responsibility/Handler.java
@@ -4,7 +4,7 @@ package behavioral.chain_of_responsibility;
  * Created by Sergei on 10.09.2015.
  */
 public interface Handler {
-    void handle(int number);
+    void handle(Long number);
 
     void setNext(Handler handler);
 }
